refactor: separate thread pool and thread handler
